feat[#37] restructured accounts directory - moved accountList and createAccount directories to components under accounts - moved getAccounts and createAccount api files to api directory under accounts
