+++ src/server/main.ts
@@ -10,7 +10,7 @@ import { account } from '@/client/types/Account';
 dotenv.config();
 
 const app = express();
-app.use(express.urlencoded());
+app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
 app.use(morgan('dev'));
